Coding challenge
Challenge description: The 'injured' state

When the player gets hit for the first time, the zombie they collided with should be destroyed, and the player will become injured. If the player gets hit while injured, the game will end.

While injured, the player has a faster movement speed, allowing them to potentially survive for a bit longer.

When the player is injured, the player sprite should be colored red.